feat: useFetch (custom) hook
Although, currently its usage  triggers an infinite loop, cause by the  `options` argument that we passed to `useFetch`. This non-primitive value (like object and array) is not stable across re-renders, so it's recreated and gets a different identity, thus will cause the `useEffect` hook to consider its dependencies have changed.